pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "mysql>=0.0.3",
    "mysql-connector>=2.2.9",
    "mysql-connector-python>=9.4.0",
    "openai-agents[litellm]>=0.2.4",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "reportlab>=4.4.3",
    "websockets>=15.0.1",
]
